backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.27
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
pydantic[email]>=2.6.1
slowapi>=0.1.9
python-multipart>=0.0.9
